+++ src/main/scala/com/github/nnnnusui/musicdsl/router/Sound.scala
@@ -10,7 +10,7 @@ trait Sound extends Directives with Output.JsonSupport {
   this: UseCase =>
   val route = (rollId: Int) =>
     get {
-      parameters("sample_rate".as[Int], "tempo".as[Double], "beat".as[Int]) { (sampleRate, tempo, beat) =>
+      parameters("sample_rate".as[Int], "tempo".as[Double], "beat".as[Double]) { (sampleRate, tempo, beat) =>
         onComplete(useCase.use(rollId, Input.Get(sampleRate, tempo, beat))) { result => complete(result) }
       }
     }
